+++ src/main/java/com/solvd/construction/model/Supplier.java
@@ -1,19 +1,19 @@
 package com.solvd.construction.model;
 
 public class Supplier {
-    private final int id;
+    private final long id;
     private String supplierName;
     private String supplierEmail;
     private Country country;
 
-    public Supplier(int id, String supplierName, String supplierEmail, Country country) {
+    public Supplier(long id, String supplierName, String supplierEmail, Country country) {
         this.id = id;
         this.supplierName = supplierName;
         this.supplierEmail = supplierEmail;
         this.country = country;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
